+++ src/components/ContentTab/StoryWebContent/StoryWebContent.vue
@@ -8,7 +8,6 @@
           v-model="name"
           for="fcb-input-name" 
           class="fcb-input-name"
-          :show-image="false"
           unstyled
           :placeholder="localize('placeholders.storyWebName')"
           :pt="{