+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - You can now create graphs tracking the relationships between your characters, locations, PCs, dangers, etc.  Right click on the new "Story webs" folder to create one and get started.
 - Main tabs (ex. entries, sessions, etc.) now remember which subtab they had open when you go back to them.  I think this is a huge quality of life improvement.
+- In the module settings, you can now configure to now show images on the description pages of different content types.  This lets you have a bit more room for the description boxes for places you don't use images.
 - Fixed a bug where deleting content in the module was still leaving stray journal entries in the setting's compendium.  It didn't really hurt anything, but it was confusing if you opened the compendium manually.  Plus they were still taking up unneeded space.
 - Fixed a bug where moving lore around between sessions, arcs, and campaigns would lose the connection to any attached journal entries.